+++ scripts/errors-last-hour.js
@@ -1,13 +1,13 @@
 require('dotenv').config()
 const log = require('skog')
-const { logSisImportErrors } = require('../lib/index')
+const logErrors = require('../src/log-errors')
 
 async function start () {
   log.info('This function fetches all SIS Import Errors that happened recently')
   const now = new Date()
   const date = new Date()
   date.setHours(now.getHours() - 1)
-  await logSisImportErrors(date, ['lms-activity-rooms'])
+  await logErrors(date, ['lms-activity-rooms'])
 }
 
 start()
